ch(tabs navigation):Add home tabs navigation  *ch(chore-app-tabs-navigation): tabs navigation setup  *fix(app routing): correct routing system
