+++ app/(app)/ActionMenu/_layout.tsx
@@ -4,8 +4,9 @@ export default function Layout() {
   return (
     <Stack>
       <Stack.Screen name="HomeScreen" />
-      <Stack.Screen name="index" />
-      <Stack.Screen name="NotificationScreen" />
+      <Stack.Screen name="index" options={{ headerShown: false }} />
+      <Stack.Screen name="NotificationScreen" options={{ headerShown: false }} />
+      <Stack.Screen name="Booking" options={{ headerShown: false }} />
     </Stack>
   );
 }
