--- app/(app)/Appointments/ReaschedualAppointment/index.tsx
@@ -1,10 +1,0 @@
-import React from 'react'
-import { View } from 'react-native'
-
-const index = () => {
-  return (
-    <View>index</View>
-  )
-}
-
-export default index
