+++ app/(app)/Appointments/index.tsx
@@ -1,0 +1,8 @@
+import React from "react";
+import { Text, View } from "react-native";
+
+const Index = () => {
+  return <Text>index</Text>;
+};
+
+export default Index;
