+++ assets/icons/AppointmentSvg.ts
@@ -1,25 +1,30 @@
-export  const AppointmentIcon = `
+export const AppointmentIcon = `
 <svg width="24" height="24" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
 <g id="Iconly/Light-Outline/Calendar">
 <g id="Calendar">
-<path id="Fill 1" fill-rule="evenodd" clip-rule="evenodd" d="M20.6665 9.9043H2.84253C2.42853 9.9043 2.09253 9.5683 2.09253 9.1543C2.09253 8.7403 2.42853 8.4043 2.84253 8.4043H20.6665C21.0805 8.4043 21.4165 8.7403 21.4165 9.1543C21.4165 9.5683 21.0805 9.9043 20.6665 9.9043Z" fill="#212121"/>
-<path id="Fill 3" fill-rule="evenodd" clip-rule="evenodd" d="M16.2013 13.8096C15.7873 13.8096 15.4473 13.4736 15.4473 13.0596C15.4473 12.6456 15.7783 12.3096 16.1923 12.3096H16.2013C16.6153 12.3096 16.9513 12.6456 16.9513 13.0596C16.9513 13.4736 16.6153 13.8096 16.2013 13.8096Z" fill="#212121"/>
-<path id="Fill 5" fill-rule="evenodd" clip-rule="evenodd" d="M11.7638 13.8096C11.3498 13.8096 11.0098 13.4736 11.0098 13.0596C11.0098 12.6456 11.3408 12.3096 11.7548 12.3096H11.7638C12.1778 12.3096 12.5138 12.6456 12.5138 13.0596C12.5138 13.4736 12.1778 13.8096 11.7638 13.8096Z" fill="#212121"/>
-<path id="Fill 7" fill-rule="evenodd" clip-rule="evenodd" d="M7.31701 13.8096C6.90301 13.8096 6.56201 13.4736 6.56201 13.0596C6.56201 12.6456 6.89401 12.3096 7.30801 12.3096H7.31701C7.73101 12.3096 8.06701 12.6456 8.06701 13.0596C8.06701 13.4736 7.73101 13.8096 7.31701 13.8096Z" fill="#212121"/>
-<path id="Fill 9" fill-rule="evenodd" clip-rule="evenodd" d="M16.2013 17.6963C15.7873 17.6963 15.4473 17.3603 15.4473 16.9463C15.4473 16.5323 15.7783 16.1963 16.1923 16.1963H16.2013C16.6153 16.1963 16.9513 16.5323 16.9513 16.9463C16.9513 17.3603 16.6153 17.6963 16.2013 17.6963Z" fill="#212121"/>
-<path id="Fill 11" fill-rule="evenodd" clip-rule="evenodd" d="M11.7638 17.6963C11.3498 17.6963 11.0098 17.3603 11.0098 16.9463C11.0098 16.5323 11.3408 16.1963 11.7548 16.1963H11.7638C12.1778 16.1963 12.5138 16.5323 12.5138 16.9463C12.5138 17.3603 12.1778 17.6963 11.7638 17.6963Z" fill="#212121"/>
-<path id="Fill 13" fill-rule="evenodd" clip-rule="evenodd" d="M7.31701 17.6963C6.90301 17.6963 6.56201 17.3603 6.56201 16.9463C6.56201 16.5323 6.89401 16.1963 7.30801 16.1963H7.31701C7.73101 16.1963 8.06701 16.5323 8.06701 16.9463C8.06701 17.3603 7.73101 17.6963 7.31701 17.6963Z" fill="#212121"/>
-<path id="Fill 15" fill-rule="evenodd" clip-rule="evenodd" d="M15.7935 5.791C15.3795 5.791 15.0435 5.455 15.0435 5.041V1.75C15.0435 1.336 15.3795 1 15.7935 1C16.2075 1 16.5435 1.336 16.5435 1.75V5.041C16.5435 5.455 16.2075 5.791 15.7935 5.791Z" fill="#212121"/>
-<path id="Fill 17" fill-rule="evenodd" clip-rule="evenodd" d="M7.71533 5.791C7.30133 5.791 6.96533 5.455 6.96533 5.041V1.75C6.96533 1.336 7.30133 1 7.71533 1C8.12933 1 8.46533 1.336 8.46533 1.75V5.041C8.46533 5.455 8.12933 5.791 7.71533 5.791Z" fill="#212121"/>
+<path id="Fill 1" fill-rule="evenodd" clip-rule="evenodd" d="M20.6665 9.9043H2.84253C2.42853 9.9043 2.09253 9.5683 2.09253 9.1543C2.09253 8.7403 2.42853 8.4043 2.84253 8.4043H20.6665C21.0805 8.4043 21.4165 8.7403 21.4165 9.1543C21.4165 9.5683 21.0805 9.9043 20.6665 9.9043Z" fill="#9E9E9E"/>
+<path id="Fill 3" fill-rule="evenodd" clip-rule="evenodd" d="M16.2013 13.8096C15.7873 13.8096 15.4473 13.4736 15.4473 13.0596C15.4473 12.6456 15.7783 12.3096 16.1923 12.3096H16.2013C16.6153 12.3096 16.9513 12.6456 16.9513 13.0596C16.9513 13.4736 16.6153 13.8096 16.2013 13.8096Z" fill="#9E9E9E"/>
+<path id="Fill 5" fill-rule="evenodd" clip-rule="evenodd" d="M11.7638 13.8096C11.3498 13.8096 11.0098 13.4736 11.0098 13.0596C11.0098 12.6456 11.3408 12.3096 11.7548 12.3096H11.7638C12.1778 12.3096 12.5138 12.6456 12.5138 13.0596C12.5138 13.4736 12.1778 13.8096 11.7638 13.8096Z" fill="#9E9E9E"/>
+<path id="Fill 7" fill-rule="evenodd" clip-rule="evenodd" d="M7.31701 13.8096C6.90301 13.8096 6.56201 13.4736 6.56201 13.0596C6.56201 12.6456 6.89401 12.3096 7.30801 12.3096H7.31701C7.73101 12.3096 8.06701 12.6456 8.06701 13.0596C8.06701 13.4736 7.73101 13.8096 7.31701 13.8096Z" fill="#9E9E9E"/>
+<path id="Fill 9" fill-rule="evenodd" clip-rule="evenodd" d="M16.2013 17.6963C15.7873 17.6963 15.4473 17.3603 15.4473 16.9463C15.4473 16.5323 15.7783 16.1963 16.1923 16.1963H16.2013C16.6153 16.1963 16.9513 16.5323 16.9513 16.9463C16.9513 17.3603 16.6153 17.6963 16.2013 17.6963Z" fill="#9E9E9E"/>
+<path id="Fill 11" fill-rule="evenodd" clip-rule="evenodd" d="M11.7638 17.6963C11.3498 17.6963 11.0098 17.3603 11.0098 16.9463C11.0098 16.5323 11.3408 16.1963 11.7548 16.1963H11.7638C12.1778 16.1963 12.5138 16.5323 12.5138 16.9463C12.5138 17.3603 12.1778 17.6963 11.7638 17.6963Z" fill="#9E9E9E"/>
+<path id="Fill 13" fill-rule="evenodd" clip-rule="evenodd" d="M7.31701 17.6963C6.90301 17.6963 6.56201 17.3603 6.56201 16.9463C6.56201 16.5323 6.89401 16.1963 7.30801 16.1963H7.31701C7.73101 16.1963 8.06701 16.5323 8.06701 16.9463C8.06701 17.3603 7.73101 17.6963 7.31701 17.6963Z" fill="#9E9E9E"/>
+<path id="Fill 15" fill-rule="evenodd" clip-rule="evenodd" d="M15.7935 5.791C15.3795 5.791 15.0435 5.455 15.0435 5.041V1.75C15.0435 1.336 15.3795 1 15.7935 1C16.2075 1 16.5435 1.336 16.5435 1.75V5.041C16.5435 5.455 16.2075 5.791 15.7935 5.791Z" fill="#9E9E9E"/>
+<path id="Fill 17" fill-rule="evenodd" clip-rule="evenodd" d="M7.71533 5.791C7.30133 5.791 6.96533 5.455 6.96533 5.041V1.75C6.96533 1.336 7.30133 1 7.71533 1C8.12933 1 8.46533 1.336 8.46533 1.75V5.041C8.46533 5.455 8.12933 5.791 7.71533 5.791Z" fill="#9E9E9E"/>
 <g id="Group 21">
 <mask id="mask0_430_8753" style="mask-type:luminance" maskUnits="userSpaceOnUse" x="2" y="2" width="20" height="21">
 <path id="Clip 20" fill-rule="evenodd" clip-rule="evenodd" d="M2 2.5791H21.5V22.5H2V2.5791Z" fill="white"/>
 </mask>
 <g mask="url(#mask0_430_8753)">
-<path id="Fill 19" fill-rule="evenodd" clip-rule="evenodd" d="M7.521 4.0791C4.928 4.0791 3.5 5.4621 3.5 7.9731V17.0221C3.5 19.5881 4.928 21.0001 7.521 21.0001H15.979C18.572 21.0001 20 19.6141 20 17.0981V7.9731C20.004 6.7381 19.672 5.7781 19.013 5.1181C18.335 4.4381 17.29 4.0791 15.988 4.0791H7.521ZM15.979 22.5001H7.521C4.116 22.5001 2 20.4011 2 17.0221V7.9731C2 4.6451 4.116 2.5791 7.521 2.5791H15.988C17.697 2.5791 19.11 3.0911 20.075 4.0581C21.012 4.9991 21.505 6.3521 21.5 7.9751V17.0981C21.5 20.4301 19.384 22.5001 15.979 22.5001Z" fill="#212121"/>
+<path id="Fill 19" fill-rule="evenodd" clip-rule="evenodd" d="M7.521 4.0791C4.928 4.0791 3.5 5.4621 3.5 7.9731V17.0221C3.5 19.5881 4.928 21.0001 7.521 21.0001H15.979C18.572 21.0001 20 19.6141 20 17.0981V7.9731C20.004 6.7381 19.672 5.7781 19.013 5.1181C18.335 4.4381 17.29 4.0791 15.988 4.0791H7.521ZM15.979 22.5001H7.521C4.116 22.5001 2 20.4011 2 17.0221V7.9731C2 4.6451 4.116 2.5791 7.521 2.5791H15.988C17.697 2.5791 19.11 3.0911 20.075 4.0581C21.012 4.9991 21.505 6.3521 21.5 7.9751V17.0981C21.5 20.4301 19.384 22.5001 15.979 22.5001Z" fill="#9E9E9E"/>
 </g>
 </g>
 </g>
 </g>
 </svg>
-`
+`;
+
+export const AppointmentFocusedIcon = `<svg width="19" height="20" viewBox="0 0 19 20" fill="none" xmlns="http://www.w3.org/2000/svg">
+<path fill-rule="evenodd" clip-rule="evenodd" d="M13.8103 0.768617L13.8113 1.51824C16.5659 1.73413 18.3856 3.61119 18.3885 6.48975L18.3994 14.9155C18.4033 18.054 16.4317 19.985 13.2712 19.99L5.5513 20C2.41061 20.004 0.414231 18.027 0.410281 14.8795L0.399421 6.55272C0.395471 3.65517 2.15094 1.78311 4.90559 1.53024L4.9046 0.780611C4.90361 0.340832 5.22943 0.00999726 5.66385 0.00999726C6.09828 0.00899776 6.4241 0.338833 6.42508 0.778612L6.42607 1.47826L12.2908 1.47027L12.2898 0.770616C12.2888 0.330837 12.6146 0.00100177 13.0491 2.26549e-06C13.4736 -0.000997234 13.8093 0.328838 13.8103 0.768617ZM1.92089 6.86157L16.869 6.84158V6.49175C16.8266 4.34283 15.7484 3.21539 13.8133 3.04748L13.8142 3.81709C13.8142 4.24688 13.4795 4.5877 13.055 4.5877C12.6206 4.5887 12.2938 4.24887 12.2938 3.81909L12.2928 3.0095L6.42804 3.01749L6.42903 3.82609C6.42903 4.25687 6.1042 4.5967 5.66978 4.5967C5.23535 4.5977 4.90855 4.25887 4.90855 3.82809L4.90756 3.05847C2.98227 3.25137 1.91694 4.38281 1.91991 6.55072L1.92089 6.86157ZM12.6393 11.4043V11.4153C12.6492 11.8751 13.0244 12.2239 13.4795 12.2139C13.9238 12.2029 14.2783 11.8221 14.2684 11.3623C14.2477 10.9225 13.8913 10.5637 13.4479 10.5647C12.9938 10.5747 12.6383 10.9445 12.6393 11.4043ZM13.4549 15.892C13.0007 15.882 12.6344 15.5032 12.6334 15.0435C12.6235 14.5837 12.9879 14.2029 13.442 14.1919H13.4519C13.9159 14.1919 14.2921 14.5707 14.2921 15.0405C14.2931 15.5102 13.9179 15.891 13.4549 15.892ZM8.57153 11.4203C8.59128 11.8801 8.96745 12.2389 9.42162 12.2189C9.86592 12.1979 10.2204 11.8181 10.2006 11.3583C10.1898 10.9085 9.82445 10.5587 9.38016 10.5597C8.92598 10.5797 8.57055 10.9605 8.57153 11.4203ZM9.42557 15.8471C8.9714 15.8671 8.59622 15.5082 8.57548 15.0485C8.57548 14.5887 8.92993 14.2089 9.3841 14.1879C9.8284 14.1869 10.1947 14.5367 10.2046 14.9855C10.2253 15.4463 9.86987 15.8261 9.42557 15.8471ZM4.50374 11.4553C4.52349 11.915 4.89966 12.2749 5.35383 12.2539C5.79813 12.2339 6.15258 11.8531 6.13185 11.3933C6.12197 10.9435 5.75666 10.5937 5.31138 10.5947C4.85721 10.6147 4.50276 10.9955 4.50374 11.4553ZM5.35778 15.8521C4.90361 15.8731 4.52843 15.5132 4.50769 15.0535C4.5067 14.5937 4.86214 14.2129 5.31631 14.1929C5.76061 14.1919 6.12691 14.5417 6.13678 14.9915C6.15752 15.4513 5.80307 15.8321 5.35778 15.8521Z" fill="#246BFD"/>
+</svg>
+`;
