+++ assets/icons/ArticleSvg.ts
@@ -1,4 +1,4 @@
-export  const ArticleIcon = `
+export const ArticleIcon = `
 <svg width="24" height="24" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
 <g id="Iconly/Light-Outline/Paper">
 <g id="Paper">
@@ -7,13 +7,18 @@ export  const ArticleIcon = `
 <path id="Clip 2" fill-rule="evenodd" clip-rule="evenodd" d="M3.00024 2.0116H20.0529V21.8651H3.00024V2.0116Z" fill="white"/>
 </mask>
 <g mask="url(#mask0_430_8283)">
-<path id="Fill 1" fill-rule="evenodd" clip-rule="evenodd" d="M7.57335 3.5116C5.91635 3.5116 4.54035 4.8536 4.50135 6.5086V17.2036C4.46435 18.9166 5.81435 20.3276 7.51035 20.3656H15.5743C17.2433 20.2966 18.5653 18.9096 18.5533 17.2096V8.3396L13.9183 3.5116H7.58535H7.57335ZM7.58535 21.8656H7.47635C4.95435 21.8086 2.94635 19.7106 3.00135 17.1876V6.4906C3.05935 4.0096 5.10835 2.0116 7.57135 2.0116H7.58835H14.2383C14.4423 2.0116 14.6373 2.0946 14.7793 2.2416L19.8443 7.5186C19.9783 7.6576 20.0533 7.8446 20.0533 8.0376V17.2036C20.0713 19.7126 18.1173 21.7626 15.6043 21.8646L7.58535 21.8656Z" fill="#212121"/>
+<path id="Fill 1" fill-rule="evenodd" clip-rule="evenodd" d="M7.57335 3.5116C5.91635 3.5116 4.54035 4.8536 4.50135 6.5086V17.2036C4.46435 18.9166 5.81435 20.3276 7.51035 20.3656H15.5743C17.2433 20.2966 18.5653 18.9096 18.5533 17.2096V8.3396L13.9183 3.5116H7.58535H7.57335ZM7.58535 21.8656H7.47635C4.95435 21.8086 2.94635 19.7106 3.00135 17.1876V6.4906C3.05935 4.0096 5.10835 2.0116 7.57135 2.0116H7.58835H14.2383C14.4423 2.0116 14.6373 2.0946 14.7793 2.2416L19.8443 7.5186C19.9783 7.6576 20.0533 7.8446 20.0533 8.0376V17.2036C20.0713 19.7126 18.1173 21.7626 15.6043 21.8646L7.58535 21.8656Z" fill="#9E9E9E"/>
 </g>
 </g>
-<path id="Fill 4" fill-rule="evenodd" clip-rule="evenodd" d="M19.2983 8.98424H16.5433C14.7133 8.97924 13.2253 7.48724 13.2253 5.65924V2.75024C13.2253 2.33624 13.5613 2.00024 13.9753 2.00024C14.3893 2.00024 14.7253 2.33624 14.7253 2.75024V5.65924C14.7253 6.66324 15.5423 7.48124 16.5453 7.48424H19.2983C19.7123 7.48424 20.0483 7.82024 20.0483 8.23424C20.0483 8.64824 19.7123 8.98424 19.2983 8.98424Z" fill="#212121"/>
-<path id="Fill 6" fill-rule="evenodd" clip-rule="evenodd" d="M13.7887 16.1084H8.38867C7.97467 16.1084 7.63867 15.7724 7.63867 15.3584C7.63867 14.9444 7.97467 14.6084 8.38867 14.6084H13.7887C14.2027 14.6084 14.5387 14.9444 14.5387 15.3584C14.5387 15.7724 14.2027 16.1084 13.7887 16.1084Z" fill="#212121"/>
-<path id="Fill 8" fill-rule="evenodd" clip-rule="evenodd" d="M11.7437 12.3563H8.3877C7.9737 12.3563 7.6377 12.0203 7.6377 11.6063C7.6377 11.1923 7.9737 10.8563 8.3877 10.8563H11.7437C12.1577 10.8563 12.4937 11.1923 12.4937 11.6063C12.4937 12.0203 12.1577 12.3563 11.7437 12.3563Z" fill="#212121"/>
+<path id="Fill 4" fill-rule="evenodd" clip-rule="evenodd" d="M19.2983 8.98424H16.5433C14.7133 8.97924 13.2253 7.48724 13.2253 5.65924V2.75024C13.2253 2.33624 13.5613 2.00024 13.9753 2.00024C14.3893 2.00024 14.7253 2.33624 14.7253 2.75024V5.65924C14.7253 6.66324 15.5423 7.48124 16.5453 7.48424H19.2983C19.7123 7.48424 20.0483 7.82024 20.0483 8.23424C20.0483 8.64824 19.7123 8.98424 19.2983 8.98424Z" fill="#9E9E9E"/>
+<path id="Fill 6" fill-rule="evenodd" clip-rule="evenodd" d="M13.7887 16.1084H8.38867C7.97467 16.1084 7.63867 15.7724 7.63867 15.3584C7.63867 14.9444 7.97467 14.6084 8.38867 14.6084H13.7887C14.2027 14.6084 14.5387 14.9444 14.5387 15.3584C14.5387 15.7724 14.2027 16.1084 13.7887 16.1084Z" fill="#9E9E9E"/>
+<path id="Fill 8" fill-rule="evenodd" clip-rule="evenodd" d="M11.7437 12.3563H8.3877C7.9737 12.3563 7.6377 12.0203 7.6377 11.6063C7.6377 11.1923 7.9737 10.8563 8.3877 10.8563H11.7437C12.1577 10.8563 12.4937 11.1923 12.4937 11.6063C12.4937 12.0203 12.1577 12.3563 11.7437 12.3563Z" fill="#9E9E9E"/>
 </g>
 </g>
 </svg>
-`
+`;
+
+export const ArticleFocusedIcon = `<svg width="25" height="24" viewBox="0 0 25 24" fill="none" xmlns="http://www.w3.org/2000/svg">
+<path fill-rule="evenodd" clip-rule="evenodd" d="M9.52535 16.39H14.9115C15.3174 16.39 15.6541 16.05 15.6541 15.64C15.6541 15.23 15.3174 14.9 14.9115 14.9H9.52535C9.11941 14.9 8.78278 15.23 8.78278 15.64C8.78278 16.05 9.11941 16.39 9.52535 16.39ZM12.8719 9.9H9.52535C9.11941 9.9 8.78278 10.24 8.78278 10.65C8.78278 11.06 9.11941 11.39 9.52535 11.39H12.8719C13.2778 11.39 13.6145 11.06 13.6145 10.65C13.6145 10.24 13.2778 9.9 12.8719 9.9ZM19.9377 9.02561C20.1704 9.02292 20.4238 9.02 20.6541 9.02C20.9016 9.02 21.0996 9.22 21.0996 9.47V17.51C21.0996 19.99 19.1095 22 16.6541 22H8.77288C6.19862 22 4.09961 19.89 4.09961 17.29V6.51C4.09961 4.03 6.09961 2 8.56496 2H13.8521C14.1095 2 14.3075 2.21 14.3075 2.46V5.68C14.3075 7.51 15.8026 9.01 17.6145 9.02C18.0377 9.02 18.4108 9.02316 18.7373 9.02593C18.9913 9.02809 19.2172 9.03 19.4164 9.03C19.5574 9.03 19.7401 9.02789 19.9377 9.02561ZM20.2107 7.566C19.3968 7.569 18.4374 7.566 17.7473 7.559C16.6523 7.559 15.7503 6.648 15.7503 5.542V2.906C15.7503 2.475 16.2681 2.261 16.5642 2.572C17.1 3.13476 17.8365 3.90834 18.5695 4.67837C19.3005 5.44632 20.0282 6.21074 20.5503 6.759C20.8394 7.062 20.6275 7.565 20.2107 7.566Z" fill="#246BFD"/>
+</svg>
+`;
