+++ assets/icons/HistorySvg.ts
@@ -1,19 +1,24 @@
-export  const HistoryIcon = `
+export const HistoryIcon = `
 <svg width="24" height="24" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
 <g id="Iconly/Light-Outline/Document">
 <g id="Document">
-<path id="Fill 1" fill-rule="evenodd" clip-rule="evenodd" d="M15.7158 16.9738H8.49585C8.08185 16.9738 7.74585 16.6378 7.74585 16.2238C7.74585 15.8098 8.08185 15.4738 8.49585 15.4738H15.7158C16.1298 15.4738 16.4658 15.8098 16.4658 16.2238C16.4658 16.6378 16.1298 16.9738 15.7158 16.9738Z" fill="#212121"/>
-<path id="Fill 3" fill-rule="evenodd" clip-rule="evenodd" d="M15.7158 12.7872H8.49585C8.08185 12.7872 7.74585 12.4512 7.74585 12.0372C7.74585 11.6232 8.08185 11.2872 8.49585 11.2872H15.7158C16.1298 11.2872 16.4658 11.6232 16.4658 12.0372C16.4658 12.4512 16.1298 12.7872 15.7158 12.7872Z" fill="#212121"/>
-<path id="Fill 5" fill-rule="evenodd" clip-rule="evenodd" d="M11.2506 8.61047H8.49561C8.08161 8.61047 7.74561 8.27447 7.74561 7.86047C7.74561 7.44647 8.08161 7.11047 8.49561 7.11047H11.2506C11.6646 7.11047 12.0006 7.44647 12.0006 7.86047C12.0006 8.27447 11.6646 8.61047 11.2506 8.61047Z" fill="#212121"/>
+<path id="Fill 1" fill-rule="evenodd" clip-rule="evenodd" d="M15.7158 16.9738H8.49585C8.08185 16.9738 7.74585 16.6378 7.74585 16.2238C7.74585 15.8098 8.08185 15.4738 8.49585 15.4738H15.7158C16.1298 15.4738 16.4658 15.8098 16.4658 16.2238C16.4658 16.6378 16.1298 16.9738 15.7158 16.9738Z" fill="#9E9E9E"/>
+<path id="Fill 3" fill-rule="evenodd" clip-rule="evenodd" d="M15.7158 12.7872H8.49585C8.08185 12.7872 7.74585 12.4512 7.74585 12.0372C7.74585 11.6232 8.08185 11.2872 8.49585 11.2872H15.7158C16.1298 11.2872 16.4658 11.6232 16.4658 12.0372C16.4658 12.4512 16.1298 12.7872 15.7158 12.7872Z" fill="#9E9E9E"/>
+<path id="Fill 5" fill-rule="evenodd" clip-rule="evenodd" d="M11.2506 8.61047H8.49561C8.08161 8.61047 7.74561 8.27447 7.74561 7.86047C7.74561 7.44647 8.08161 7.11047 8.49561 7.11047H11.2506C11.6646 7.11047 12.0006 7.44647 12.0006 7.86047C12.0006 8.27447 11.6646 8.61047 11.2506 8.61047Z" fill="#9E9E9E"/>
 <g id="Group 9">
 <mask id="mask0_430_8874" style="mask-type:luminance" maskUnits="userSpaceOnUse" x="3" y="2" width="19" height="20">
 <path id="Clip 8" fill-rule="evenodd" clip-rule="evenodd" d="M3 2.00024H21.1647V21.91H3V2.00024Z" fill="white"/>
 </mask>
 <g mask="url(#mask0_430_8874)">
-<path id="Fill 7" fill-rule="evenodd" clip-rule="evenodd" d="M15.9088 3.5L8.21976 3.504C5.89176 3.518 4.49976 4.958 4.49976 7.357V16.553C4.49976 18.968 5.90476 20.41 8.25576 20.41L15.9448 20.407C18.2728 20.393 19.6648 18.951 19.6648 16.553V7.357C19.6648 4.942 18.2608 3.5 15.9088 3.5ZM8.25676 21.91C5.11276 21.91 2.99976 19.757 2.99976 16.553V7.357C2.99976 4.124 5.04676 2.023 8.21476 2.004L15.9078 2H15.9088C19.0528 2 21.1648 4.153 21.1648 7.357V16.553C21.1648 19.785 19.1178 21.887 15.9498 21.907L8.25676 21.91Z" fill="#212121"/>
+<path id="Fill 7" fill-rule="evenodd" clip-rule="evenodd" d="M15.9088 3.5L8.21976 3.504C5.89176 3.518 4.49976 4.958 4.49976 7.357V16.553C4.49976 18.968 5.90476 20.41 8.25576 20.41L15.9448 20.407C18.2728 20.393 19.6648 18.951 19.6648 16.553V7.357C19.6648 4.942 18.2608 3.5 15.9088 3.5ZM8.25676 21.91C5.11276 21.91 2.99976 19.757 2.99976 16.553V7.357C2.99976 4.124 5.04676 2.023 8.21476 2.004L15.9078 2H15.9088C19.0528 2 21.1648 4.153 21.1648 7.357V16.553C21.1648 19.785 19.1178 21.887 15.9498 21.907L8.25676 21.91Z" fill="#9E9E9E"/>
 </g>
 </g>
 </g>
 </g>
 </svg>
-`
+`;
+
+export const HistoryFocusedIcon = `<svg width="24" height="24" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
+<path fill-rule="evenodd" clip-rule="evenodd" d="M7.81 2H16.191C19.28 2 21 3.78 21 6.83V17.16C21 20.26 19.28 22 16.191 22H7.81C4.77 22 3 20.26 3 17.16V6.83C3 3.78 4.77 2 7.81 2ZM8.08 6.66V6.65H11.069C11.5 6.65 11.85 7 11.85 7.429C11.85 7.87 11.5 8.22 11.069 8.22H8.08C7.649 8.22 7.3 7.87 7.3 7.44C7.3 7.01 7.649 6.66 8.08 6.66ZM8.08 12.74H15.92C16.35 12.74 16.7 12.39 16.7 11.96C16.7 11.53 16.35 11.179 15.92 11.179H8.08C7.649 11.179 7.3 11.53 7.3 11.96C7.3 12.39 7.649 12.74 8.08 12.74ZM8.08 17.31H15.92C16.319 17.27 16.62 16.929 16.62 16.53C16.62 16.12 16.319 15.78 15.92 15.74H8.08C7.78 15.71 7.49 15.85 7.33 16.11C7.17 16.36 7.17 16.69 7.33 16.95C7.49 17.2 7.78 17.35 8.08 17.31Z" fill="#246BFD"/>
+</svg>
+`;
