+++ assets/icons/HomeSvg.ts
@@ -1,5 +1,4 @@
-
-  export  const HomeIcon = `
+export const HomeIcon = `
     <svg width="24" height="24" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
       <g id="Iconly/Light-Outline/Home">
         <g id="Home">
@@ -7,10 +6,14 @@
             <path id="Clip 2" fill-rule="evenodd" clip-rule="evenodd" d="M2 1.00024H22.4998V22.5052H2V1.00024Z" fill="white"/>
           </mask>
           <g mask="url(#mask0_430_8251)">
-            <path id="Fill 1" fill-rule="evenodd" clip-rule="evenodd" d="M13.7168 15.2914C14.9208 15.2914 15.9008 16.2644 15.9008 17.4604V20.5364C15.9008 20.7934 16.1068 20.9994 16.3708 21.0054H18.2768C19.7788 21.0054 20.9998 19.7994 20.9998 18.3174V9.59337C20.9928 9.08337 20.7498 8.60337 20.3328 8.28437L13.7398 3.02637C12.8548 2.32537 11.6168 2.32537 10.7288 3.02837L4.18076 8.28237C3.74776 8.61137 3.50476 9.09137 3.49976 9.61037V18.3174C3.49976 19.7994 4.72076 21.0054 6.22276 21.0054H8.14676C8.41776 21.0054 8.63776 20.7904 8.63776 20.5264C8.63776 20.4684 8.64476 20.4104 8.65676 20.3554V17.4604C8.65676 16.2714 9.63076 15.2994 10.8258 15.2914H13.7168ZM18.2768 22.5054H16.3528C15.2508 22.4794 14.4008 21.6144 14.4008 20.5364V17.4604C14.4008 17.0914 14.0938 16.7914 13.7168 16.7914H10.8308C10.4618 16.7934 10.1568 17.0944 10.1568 17.4604V20.5264C10.1568 20.6014 10.1468 20.6734 10.1258 20.7414C10.0178 21.7314 9.17176 22.5054 8.14676 22.5054H6.22276C3.89376 22.5054 1.99976 20.6264 1.99976 18.3174V9.60337C2.00976 8.60937 2.46776 7.69937 3.25876 7.10037L9.79376 1.85537C11.2328 0.715372 13.2378 0.715372 14.6738 1.85337L21.2558 7.10337C22.0288 7.69237 22.4868 8.60037 22.4998 9.58237V18.3174C22.4998 20.6264 20.6058 22.5054 18.2768 22.5054Z" fill="#212121"/>
+            <path id="Fill 1" fill-rule="evenodd" clip-rule="evenodd" d="M13.7168 15.2914C14.9208 15.2914 15.9008 16.2644 15.9008 17.4604V20.5364C15.9008 20.7934 16.1068 20.9994 16.3708 21.0054H18.2768C19.7788 21.0054 20.9998 19.7994 20.9998 18.3174V9.59337C20.9928 9.08337 20.7498 8.60337 20.3328 8.28437L13.7398 3.02637C12.8548 2.32537 11.6168 2.32537 10.7288 3.02837L4.18076 8.28237C3.74776 8.61137 3.50476 9.09137 3.49976 9.61037V18.3174C3.49976 19.7994 4.72076 21.0054 6.22276 21.0054H8.14676C8.41776 21.0054 8.63776 20.7904 8.63776 20.5264C8.63776 20.4684 8.64476 20.4104 8.65676 20.3554V17.4604C8.65676 16.2714 9.63076 15.2994 10.8258 15.2914H13.7168ZM18.2768 22.5054H16.3528C15.2508 22.4794 14.4008 21.6144 14.4008 20.5364V17.4604C14.4008 17.0914 14.0938 16.7914 13.7168 16.7914H10.8308C10.4618 16.7934 10.1568 17.0944 10.1568 17.4604V20.5264C10.1568 20.6014 10.1468 20.6734 10.1258 20.7414C10.0178 21.7314 9.17176 22.5054 8.14676 22.5054H6.22276C3.89376 22.5054 1.99976 20.6264 1.99976 18.3174V9.60337C2.00976 8.60937 2.46776 7.69937 3.25876 7.10037L9.79376 1.85537C11.2328 0.715372 13.2378 0.715372 14.6738 1.85337L21.2558 7.10337C22.0288 7.69237 22.4868 8.60037 22.4998 9.58237V18.3174C22.4998 20.6264 20.6058 22.5054 18.2768 22.5054Z" fill="#9E9E9E"/>
           </g>
         </g>
       </g>
     </svg>
-  `
+  `;
 
+export const HomeFocusedIcon = `<svg width="25" height="24" viewBox="0 0 25 24" fill="none" xmlns="http://www.w3.org/2000/svg">
+  <path d="M9.93458 20.7733V17.7156C9.93458 16.9351 10.572 16.3023 11.3582 16.3023H14.2324C14.61 16.3023 14.9721 16.4512 15.2391 16.7163C15.5061 16.9813 15.6561 17.3408 15.6561 17.7156V20.7733C15.6537 21.0978 15.7819 21.4099 16.0122 21.6402C16.2425 21.8705 16.5559 22 16.8827 22H18.8436C19.7594 22.0023 20.6386 21.6428 21.287 21.0008C21.9354 20.3588 22.2998 19.487 22.2998 18.5778V9.86686C22.2998 9.13246 21.9719 8.43584 21.4044 7.96467L14.7338 2.67587C13.5735 1.74856 11.9109 1.7785 10.7852 2.74698L4.26682 7.96467C3.67255 8.42195 3.31736 9.12064 3.2998 9.86686V18.5689C3.2998 20.4639 4.84719 22 6.75598 22H8.67209C9.35103 22 9.9028 21.4562 9.90772 20.7822L9.93458 20.7733Z" fill="#246BFD"/>
+  </svg>
+  `;
